server/requirements.txt
pythonql3==0.9.61
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.6.0
decorator==4.4.0
Deprecated==1.2.6
ecdsa==0.13.2
Flask==1.1.1
Flask-Cors==3.0.6
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
grlc==1.3.0
gunicorn==19.6.0
html5lib==1.0.1
idna==2.7
inflection==0.3.1
isodate==0.5.4
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
keepalive==0.5
MarkupSafe==0.23
obasparql==2.1.1
openapi-spec-validator==0.2.8
pathlib==1.0.1
ply==3.11
pyaml==18.11.0
pyasn1==0.4.7
PyGithub==1.43.5
PyJWT==1.7.1
PyLD==1.0.5
pyparsing==2.0.7
python-dateutil==2.8.0
python-jose==3.0.1
PyYAML==5.3
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.20.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
SPARQLTransformer==1.6.10
SPARQLWrapper==1.8.2
swagger-ui-bundle==0.0.5
typing==3.7.4.1
urllib3==1.24.3
validators==0.14.0
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
